+++ app/src/main/java/app/revanced/manager/ui/screen/SelectedAppInfoScreen.kt
@@ -389,6 +389,8 @@ fun SelectedAppInfoScreen(
         ) {
             AppInfo(
                 appInfo = vm.selectedAppInfo,
+                labelOverride = vm.selectedAppInfoLabelOverride,
+                iconOverride = vm.selectedAppInfoIconOverride,
                 placeholderLabel = packageName,
                 placeholderMetaLines = 2
             ) {